.cosmetics
git-svn-id: svn://ultimatepp.org/upp/trunk@8093 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Painter/SvgUtil.cpp
@@ -205,7 +205,7 @@ Array<double> GetTransformArgs(String str, const char *command) {
 	if ((pos = str.Find(command)) < 0) 
 		return args;
 	
-	pos += strlen(command);
+	pos += (int)strlen(command);
 	pos = 1 + str.Find('(', pos);
 	
 	while (true) {
